Score
CheckDistanceFromStart.cs is a class on the player game object. It calculate the distance between a starting position (empty game object) and to where the player is now.
DisplayDistance.cs is a class that will display the distance (rounded) on the UI using the "DistTraveld" property.

Added TMPro to the project